Move createBadgeDataUrl() into separate utils file
